Refactor `CutSet.describe` to enable parallel statistics computation
